# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meeting AI Visualizer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VisualizeRequest {
  /** The full normalized transcript text to visualize */
  transcript: string;
  /**
   * Previous visualization HTML for incremental updates
   * @nullable
   */
  previousHtml?: string | null;
  /**
   * Room ID to broadcast the visualization to
   * @nullable
   */
  roomId?: string | null;
  /**
   * Name of the participant triggering visualization
   * @nullable
   */
  speakerName?: string | null;
}

export interface DeepgramTokenResponse {
  /** Deepgram API key or short-lived token */
  key: string;
}

export interface SegmentRequest {
  /** The room to post to */
  roomId: string;
  /** Name of the speaker */
  speakerName: string;
  /** The transcribed text segment */
  text: string;
  /** Unix timestamp (ms) of the segment */
  timestamp: number;
  /** Whether the segment is final (not interim) */
  isFinal: boolean;
}

export interface SegmentResponse {
  ok: boolean;
  segmentId: string;
}

export interface TranscriptSegment {
  id: string;
  speakerName: string;
  text: string;
  timestamp: number;
  isFinal: boolean;
}

export interface MeetingRecord {
  id: string;
  roomId: string;
  createdAt: number;
  segments: TranscriptSegment[];
  /** @nullable */
  lastVisualization?: string | null;
}

export interface HistoryResponse {
  meetings: MeetingRecord[];
}