# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Meeting AI Visualizer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visualize:
    post:
      operationId: visualize
      tags: [visualize]
      summary: Generate AI visualization from transcript
      description: Streams HTML visualization from Claude as SSE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisualizeRequest"
      responses:
        "200":
          description: SSE stream of visualization chunks
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    VisualizeRequest:
      type: object
      properties:
        transcript:
          type: string
          description: The full normalized transcript text to visualize
        previousHtml:
          type: ["string", "null"]
          description: Previous visualization HTML for incremental updates
        roomId:
          type: ["string", "null"]
          description: Room ID to broadcast the visualization to
        speakerName:
          type: ["string", "null"]
          description: Name of the participant triggering visualization
      required:
        - transcript
...