# lib/api-zod/src/generated/types/visualizeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Meeting AI Visualizer API
 * OpenAPI spec version: 0.1.0
 */

export interface VisualizeRequest {
  /** The full normalized transcript text to visualize */
  transcript: string;
  /**
   * Previous visualization HTML for incremental updates
   * @nullable
   */
  previousHtml?: string | null;
  /**
   * Room ID to broadcast the visualization to
   * @nullable
   */
  roomId?: string | null;
  /**
   * Name of the participant triggering visualization
   * @nullable
   */
  speakerName?: string | null;
}